fix(make): defines cmd targets as PHONY (#36)
Each target which is not representing an actual file should be
marked as `.PHONY`.

Otherwise running e.g. `make build` twice would result in subsequent
execution being skipped with:

```sh
$ make build
make: 'build' is up to date.
```

which is not desired (coincidently we have both `build` and `test`
folders in this project).

This change ensures all fake/phony targets are marked respectively and
so they are executed each time.