feat(build): configurable container builder (#65)
Our build process relies on `docker` client to build and push images. In
most cases it is all we need, but we can easily open it up for other
container tools to be used instead, for example `podman`.

To make it possible, `CONTAINER_CLI` variable is introduced that
can be configured when running the build. Named is aligned with the Istio
project.

This can be used in a few different ways during the build process:

```sh
make docker -e CONTAINER_CLI=podman
CONTAINER_CLI=podman make docker
export CONTAINER_CLI=podman && make docker
```