feat(ci): ensures license headers are added (#90)
Since the `make add-license` target is not part of the default Makefile target,
there's no guarantee it will be run before a PR is opened or updated. To address
this, a CI build job ensures licenses are present where needed.

The `linters/check-license` GitHub job verifies that every relevant file includes
the required license headers. Otherwise job fails and lists files both in the logs
of the job as well as in the summary.