feat(adsc): propagates ctx when creating gRPC stream (#88)
To gracefully handle cancellation we can pass main context down to gRPC
stream creation instead of creating background context.

This change introduces ability to pass context.Context instance to
relevant methods of ADS Client. It also removes a need for keeping
client instance as part of the struct, as this is only needed to create
the stream and is never used afterwads.

In addition, any error occuring when sending initial discovery requests
is not logged.