fix(make): ensures tools installation for e2e tests (#96)
E2E tests rely on external tools such as `helm`. With the introduction
of `tools` target we are able to provide those binaries with the version
we need, but running `make e2e` on freshly cloned project can fail, as
ensuring that right tools are on the `PATH` is not part of it.

This change adds `tools` as dependent target for `e2e` and makes
required binaries part of used `PATH` so test code can invoked needed
commands.