fix(cmd): wait for root context
We shouldn't wait indefinitely by using `select {}`, but until the main
context is Done. It ensures we exit the process cleanly in
case of termination or cancel as so kubelet can restart the controller
instead of keeping it running in a broken state (i.e. reconiller will stop
working when ctx is Done).