fix(e2e): aligns echo app version with Istio (#105)
When we do not set `--istio.test.tag` when running tests (which can
happen if try from the IDE), Echo App uses `latest` tag, as we do not
propogate default `istioVersion` we have defined in the test setup.

As a consequence tests are failing, because `app:latest` is 5y old and
does not support flags defined for the container in Istio Test Framework
template.

Additionally `istioVersion` variable has been aligned to what's defined
as default in `Makefile`.