chore(make): simplifies build target (#114)
Aims to make `build` as lean as possible with the only goal to produce
the binary. At the same time moving license and import fixes to chain of
default targets.

Without this change there's always license added after regenarting
protobuf go stubs. When using tools such as air-verse/air to rebuild the
project on code changes this leads to unnecessary build loops, as one
change triggers `make build` process which leads to watch reacting on
licenses added to source files and re-triggering the build.

> [!NOTE]
> This can be obviously solved by simplifying rebuild command in the tool itself by
> swapping it from `make build` to `go build` but I thought we could
> simplify it on the `Makefile` level.