chore(docs): simplifies spire example
There are a few minor shortcomings how the current spire example is
structured.

Each time bundle propagation is executed it appends to the existing file
which is a template and under the source control. This PR reworks it to
apply changes in place.

In addition `trust-bundle-federation.yaml`s are places in a
counter-intuitive way, where the "east" one is in the "west" folder and
the other way around. If we add third cluster, we would end up with 6 of
those files, which in fact are templates. This PR changes it to have one
template with cluster name being replaced on the fly.

The <placeholders> have been renamed to ${VAR} so the files can be also
processed by tools like `envsubst` easily.

Helm installation of Spire has `--wait` flag now to prevent from
federating bundles too early which might lead to missing bundles in one
of the clusters.