feat(asdc): adds peer name to the logger (#118)
To easily distinguish source of the log message we can enhance loggers
with key-value pairs which are going to be appeneded at the end of it.
This can be especially useful with the upcoming multi-peer support.
Adding peer name to discovery client logs will help to quickly identify
the origin of it.

This PR moves log from global variable to part of Discover Service
Client and enhances it with peer name.