fix(fds): simplifies sync logic (#134)
My understanding of the current client-server flow is as follows:

- client connects to the server
- it requests push of predefined types (ExportedSvc) as initial push
  request - SotW
- on subscriber(client) connect, server does another SotW push

Not only we are sending the same data twice, but we also need to make
sure that handlers we define are aligned with initial push requests.
If we add new handler, we need to explictly define it as part of
initial push. It also needs to be defined (again) in onSubscribe
on the FDS server side.

This PR moves the initial sync responsibility to the client, ensuring
each handler automatically requests the data it needs. This avoids
redundant pushes and reduces code duplication.

Signed-off-by: bartoszmajsak <bartosz.majsak@gmail.com>

Signed-off-by: bartoszmajsak <bartosz.majsak@gmail.com>