feat(testing): introduces k8s envtest skeleton
This PR provides a skeleton for testing controller using K8s EnvTest
testing harness. This approach enables us to run reasonably fast
integration tests interacting with kube-apiserver.

This way we can have automated feedback loop for verifying parts of the
system behaviour without a need to run entire e2e suite. This is
especially handy right now, as we are fleshing out controller-runtime
implementation and have e2e tests for the legacy implementation.

Besides simple test this commit brings a facade that simplifies
configuration and use Kubernetes Env Test [1] library. The main goal of
this enhancement is to reduce boilerplate of test suite
bootstrap code by hiding the details of EnvTest setup and
wiring the controller-under-test.

Additionally, the facade struct gives an unified way of manipulating
cluster under test as it is composed of k8s client.Client and custom
Cleaner (see its godoc for details why it is needed) struct.

The tests are written in Ginkgo [2] and use Gomega matchers [3].
These tools used in many k8s projects, especially controllers, as
controller-runtime provides basic template for tests based on those
libraries. Ginkgo provides a structured and expressive way to write tests.

I am open to rework it to any other testing approach moving forward if
we feel this is not comfortable approach. This PR is based on the work I
did in other OSS projects.

[1] https://book.kubebuilder.io/reference/envtest
[2] https://onsi.github.io/ginkgo
[3] https://onsi.github.io/gomega/

Fixes #146

Signed-off-by: bartoszmajsak <bartosz.majsak@gmail.com>