feat(build): replaces goimports with gci (#160)
While working on #159 my IDE moved local imports around
and `goimports` didn't reformat it properly. It would be good to have
clear rules for grouping and separating libraries we rely on that are
applied deterministically. It seems that `goimports` is fairly limited
and can handle only sorting and basic grouping. This can lead to noise
in PRs which we shouldn't bother about.

This PR changes `fix-imports` target to use `gci` tool instead. This
tool offers more granular control over group definitions so we can
ensure consistency across the codebase.

> [!NOTE]
> Even though `gci` is part of `golangci-lint` I deliberately added it
> as a standalone tool. Introducing the latter is a bigger task which I
> hope I can soon contribute.

Signed-off-by: bartoszmajsak <bartosz.majsak@gmail.com>

Signed-off-by: bartoszmajsak <bartosz.majsak@gmail.com>