chore(build): reorganizes Makefile (#162)
As our build chain is growing it might be helpful to split Makefile to
several dedicated files which we can compose in the main `Makefile`.

This will help with identifying revelant targets quicker and group
together certain aspects of build process for easier maintenance.

This PR introduces `Makefile.tooling.mk`, where all targets for fetching
tools are now located.

### Additional fixes

- `make help` has a correct pattern matching and shows targets that
  contain numbers in them (i.e. `e2e`)
- test targets are grouped together
- `OUT_DIR` is prepended with `PROJECT_DIR`

Signed-off-by: bartoszmajsak <bartosz.majsak@gmail.com>

Signed-off-by: bartoszmajsak <bartosz.majsak@gmail.com>