feat(ctrl): predicate on finalizer change (#165)
`finalizer.Handler` introduced in #164 is expected to retrigger
reconcile when finalizer is added for the first time. This ensures that
given controller adds its cleanup hook to the resource as soon as
possible.

However adding finalizer does not increase generation of the resource,
therefore using
`WithEventFilter(predicate.GenerationChangedPredicate{})`
will not cause another reconcile.

This PR brings `FinalizerChanged` predicate which can be combined:

```golang
WithEventFilter(
	predicate.Or(
		predicate.GenerationChangedPredicate{},
		FinalizerChanged(),
	),
).
```

and it will trigger reconciliation on the finalizer change.

Signed-off-by: bartoszmajsak <bartosz.majsak@gmail.com>

Signed-off-by: bartoszmajsak <bartosz.majsak@gmail.com>