docs: replace `trustDomainAliases` with `caCertificates[].trustDomains` in examples for SPIRE (#181)
Using `trustDomainAliases` is not the right solution when federated
services from different trust domains may have overlapping service
account and namespace names. Then authorization policies defined for a
single trust domain are automatically generated for all trust domain
aliases, and this may be not desired in multi-mesh deployments, so we
should avoid using this solution.

We also use `trustDomainAliases` in e2e tests, but that can be fixed in
a follow-up.

---------

Signed-off-by: Jacek Ewertowski <jacek.ewertowski1@gmail.com>